Create SamBedToCustomBed.py
This script takes a bed file resulting from SamMismatchesToBed.py and converts it to a more general bed format suitable for mutperiod. Also, data can be filtered based on read length, number of mismatches, the sequence of the mismatch, and the position of the mismatch within the read. Both single-base and tandem mismatches can be selected for this way.