Create FilterOnExpandedContext.py
This script filters out reads based on the sequence after the cut site in their expanded context. For example, reads with a 3' post-cut-site TGG could be filtered out. Right now, this script also requires that a companion unexpanded file be present in the same directory, which will be filtered in tandem. This may become optional in the future if the need arises.